functions/get_selected_plots_local/requirements.txt
pip==20.0.2
setuptools==46.1.3
oauth2client==4.1.3
google-api-python-client==1.8.0
grpcio==1.30.0
google-cloud-firestore==1.8.1
